PROJECT 3.1
Business Scenario  Rollin Ford, the regional manager of a retail chain, is in process of re-designing distribution and production strategies. He wants to look at profit for each state along with names of managers. He also wants to see the number of products returned by customers in each state. Combine the data that is stored across different tables and create a visualization to satisfy the above requirement. In order to enhance the performance, create an extract for this data source. Rollin wants refresh to be happen incrementally based on date.   
Using the view, answer the following questions:  
• Which is the least profitable state?   • Customers from which state have returned the highest number of products?